+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- New config: `angular` (TODO).
+- Set `allowWithDecorator: true` for [`@typescript-eslint/no-extraneous-class`](https://typescript-eslint.io/rules/no-extraneous-class) rule, mostly to avoid unfixable reports in Angular projects.
+
 ## 0.6.2
 
 - Revert `eslint-import-resolver-typescript` to v3.7.0 due to [this issue](https://github.com/import-js/eslint-import-resolver-typescript/issues/364).